Remove support for the legacy etcd provider as of k8s 1.18
